# backends/internal/gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Ecommerce API
    description: Ecommerce API Design to Programming The Internet Class at BIU University
    contact:
        name: Adrison Gomez
        email: info@adrisongomez.me
        url: https://github.com/adrisongomez
    version: 0.0.1
host: localhost:3030
consumes:
    - application/json
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /api/products:
        get:
            tags:
                - svc-products
            summary: listProduct svc-products
            description: List products
            operationId: svc-products#listProduct
            parameters:
                - name: pageSize
                  in: query
                  description: Record per page
                  required: false
                  type: integer
                  default: 10
                  maximum: 100
                  minimum: 10
                - name: after
                  in: query
                  description: Start listing after this resource
                  required: false
                  type: integer
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/ProductsList'
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/SvcProductsListProductBadRequestResponseBody'
            schemes:
                - http
        post:
            tags:
                - svc-products
            summary: createProduct svc-products
            description: Create a new product
            operationId: svc-products#createProduct
            produces:
                - application/json
            parameters:
                - name: CreateProductRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/ProductInput'
                    required:
                        - title
                        - description
                        - tags
                        - vendorId
            responses:
                "201":
                    description: Created response.
                    schema:
                        $ref: '#/definitions/Product'
                "409":
                    description: Conflict response.
                    schema:
                        $ref: '#/definitions/SvcProductsCreateProductConflictResponseBody'
            schemes:
                - http
    /api/products/{productId}:
        get:
            tags:
                - svc-products
            summary: getProductById svc-products
            description: Get a product by its id
            operationId: svc-products#getProductById
            produces:
                - application/json
            parameters:
                - name: productId
                  in: path
                  description: Unique product identifier
                  required: true
                  type: integer
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/Product'
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/SvcProductsGetProductByIDNotFoundResponseBody'
            schemes:
                - http
        put:
            tags:
                - svc-products
            summary: deleteProductById svc-products
            description: Create a new product
            operationId: svc-products#deleteProductById
            parameters:
                - name: productId
                  in: path
                  description: Unique product identifier
                  required: true
                  type: integer
            responses:
                "200":
                    description: OK response.
                    schema:
                        type: boolean
                "404":
                    description: Not Found response.
                    schema:
                        $ref: '#/definitions/SvcProductsDeleteProductByIDNotFoundResponseBody'
            schemes:
                - http
definitions:
...
    ProductInput:
        title: ProductInput
        type: object
        properties:
            description:
                type: string
                description: Product description
                example: Vero temporibus adipisci.
            handle:
                type: string
                description: Last part of the url which use to idepntify the user
                example: Nihil asperiores eum animi sit ut.
            medias:
                type: array
                items:
                    $ref: '#/definitions/ProductMediaInput'
                example:
                    - alt: Sed quis.
                      mediaId: Tenetur qui consequuntur.
                      sortNumber: 2548002682993232380
                    - alt: Sed quis.
                      mediaId: Tenetur qui consequuntur.
                      sortNumber: 2548002682993232380
            status:
                type: string
                description: Product's status
                example: ACTIVE
                enum:
                    - ACTIVE
                    - DRAFT
            tags:
                type: array
                items:
                    type: string
                    example: Sunt nemo perferendis architecto.
                description: Product tags
                example:
                    - Perspiciatis eos itaque repudiandae sit veritatis blanditiis.
                    - Aut accusantium nemo sunt distinctio.
                    - Fugit aut.
            title:
                type: string
                description: Title's product
                example: Eveniet voluptas molestiae consequatur facilis.
            variants:
                type: array
                items:
                    $ref: '#/definitions/ProductVariantInput'
                description: Product variants
                example:
                    - colorHex: Molestiae et soluta animi aliquam eaque.
                      colorName: Veritatis impedit sequi tenetur numquam ad.
                      price: 9013296666552782995
                    - colorHex: Molestiae et soluta animi aliquam eaque.
                      colorName: Veritatis impedit sequi tenetur numquam ad.
                      price: 9013296666552782995
                    - colorHex: Molestiae et soluta animi aliquam eaque.
                      colorName: Veritatis impedit sequi tenetur numquam ad.
                      price: 9013296666552782995
                    - colorHex: Molestiae et soluta animi aliquam eaque.
                      colorName: Veritatis impedit sequi tenetur numquam ad.
                      price: 9013296666552782995
            vendorId:
                type: string
                description: Vendor's product
                example: Est et nostrum sit fugit dolores est.
        example:
            description: Debitis et impedit tempora ullam.
            handle: Animi et necessitatibus a aut.
            medias:
                - alt: Sed quis.
                  mediaId: Tenetur qui consequuntur.
                  sortNumber: 2548002682993232380
                - alt: Sed quis.
                  mediaId: Tenetur qui consequuntur.
                  sortNumber: 2548002682993232380
            status: ACTIVE
            tags:
                - Amet occaecati quia nam qui.
                - Numquam qui ut rerum cupiditate et.
                - Perspiciatis commodi asperiores sit.
            title: Accusantium magnam.
            variants:
                - colorHex: Molestiae et soluta animi aliquam eaque.
                  colorName: Veritatis impedit sequi tenetur numquam ad.
                  price: 9013296666552782995
                - colorHex: Molestiae et soluta animi aliquam eaque.
                  colorName: Veritatis impedit sequi tenetur numquam ad.
                  price: 9013296666552782995
            vendorId: Et fugiat enim rem quos.
        required:
            - title
            - description
            - tags
            - vendorId
...
    ProductsList:
        title: 'Mediatype identifier: application/vnd.products-list+json; view=default'
        type: object
        properties:
            data:
                $ref: '#/definitions/ProductResponseBodyCollection'
            pageInfo:
                $ref: '#/definitions/PageInfoResponseBody'
        description: ListProductResponseBody result type (default view)
        example:
            data:
                - description: Sed molestiae molestiae rerum aliquid a.
                  handle: Totam at magni eaque quia culpa.
                  id: 10
                  medias:
                    - id: 10
                      mediaType: Deleniti et itaque sapiente.
                      url: Qui velit facere sed qui qui.
                    - id: 10
                      mediaType: Deleniti et itaque sapiente.
                      url: Qui velit facere sed qui qui.
                    - id: 10
                      mediaType: Deleniti et itaque sapiente.
                      url: Qui velit facere sed qui qui.
                  status: DRAFT
                  tags:
                    - Deserunt quis suscipit qui.
                    - Quod facilis quod impedit et tempora libero.
                    - Laboriosam eum dicta.
                    - Exercitationem sit voluptatem et quia modi dolore.
                  title: Impedit optio.
                  variants:
                    - colorHex: Eos praesentium ipsa.
                      colorName: Deserunt dolores itaque.
                      featureMediaId: Nihil error dolorem error voluptate molestiae.
                      id: 10
                      price: 1759365473590346564
                      productId: Rerum molestiae omnis adipisci quidem.
                    - colorHex: Eos praesentium ipsa.
                      colorName: Deserunt dolores itaque.
                      featureMediaId: Nihil error dolorem error voluptate molestiae.
                      id: 10
                      price: 1759365473590346564
                      productId: Rerum molestiae omnis adipisci quidem.
                  vendor:
                    id: 10
                    name: Nostrum est a error.
                - description: Sed molestiae molestiae rerum aliquid a.
                  handle: Totam at magni eaque quia culpa.
                  id: 10
                  medias:
                    - id: 10
                      mediaType: Deleniti et itaque sapiente.
                      url: Qui velit facere sed qui qui.
                    - id: 10
                      mediaType: Deleniti et itaque sapiente.
                      url: Qui velit facere sed qui qui.
                    - id: 10
                      mediaType: Deleniti et itaque sapiente.
                      url: Qui velit facere sed qui qui.
                  status: DRAFT
                  tags:
                    - Deserunt quis suscipit qui.
                    - Quod facilis quod impedit et tempora libero.
                    - Laboriosam eum dicta.
                    - Exercitationem sit voluptatem et quia modi dolore.
                  title: Impedit optio.
                  variants:
                    - colorHex: Eos praesentium ipsa.
                      colorName: Deserunt dolores itaque.
                      featureMediaId: Nihil error dolorem error voluptate molestiae.
                      id: 10
                      price: 1759365473590346564
                      productId: Rerum molestiae omnis adipisci quidem.
                    - colorHex: Eos praesentium ipsa.
                      colorName: Deserunt dolores itaque.
                      featureMediaId: Nihil error dolorem error voluptate molestiae.
                      id: 10
                      price: 1759365473590346564
                      productId: Rerum molestiae omnis adipisci quidem.
                  vendor:
                    id: 10
                    name: Nostrum est a error.
                - description: Sed molestiae molestiae rerum aliquid a.
                  handle: Totam at magni eaque quia culpa.
                  id: 10
                  medias:
                    - id: 10
                      mediaType: Deleniti et itaque sapiente.
                      url: Qui velit facere sed qui qui.
                    - id: 10
                      mediaType: Deleniti et itaque sapiente.
                      url: Qui velit facere sed qui qui.
                    - id: 10
                      mediaType: Deleniti et itaque sapiente.
                      url: Qui velit facere sed qui qui.
                  status: DRAFT
                  tags:
                    - Deserunt quis suscipit qui.
                    - Quod facilis quod impedit et tempora libero.
                    - Laboriosam eum dicta.
                    - Exercitationem sit voluptatem et quia modi dolore.
                  title: Impedit optio.
                  variants:
                    - colorHex: Eos praesentium ipsa.
                      colorName: Deserunt dolores itaque.
                      featureMediaId: Nihil error dolorem error voluptate molestiae.
                      id: 10
                      price: 1759365473590346564
                      productId: Rerum molestiae omnis adipisci quidem.
                    - colorHex: Eos praesentium ipsa.
                      colorName: Deserunt dolores itaque.
                      featureMediaId: Nihil error dolorem error voluptate molestiae.
                      id: 10
                      price: 1759365473590346564
                      productId: Rerum molestiae omnis adipisci quidem.
                  vendor:
                    id: 10
                    name: Nostrum est a error.
            pageInfo:
                endCursor: Voluptas debitis neque quis.
                hasMore: false
                startCursor: Facilis repellat praesentium voluptas.
                totalResource: 3707296898274386968
    SvcProductsCreateProductConflictResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: false
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: true
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: true
        description: createProduct_Conflict_response_body result type (default view)
        example:
            fault: true
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: false
            timeout: true
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
...
    SvcProductsListProductBadRequestResponseBody:
        title: 'Mediatype identifier: application/vnd.goa.error; view=default'
        type: object
        properties:
            fault:
                type: boolean
                description: Is the error a server-side fault?
                example: false
            id:
                type: string
                description: ID is a unique identifier for this particular occurrence of the problem.
                example: 123abc
            message:
                type: string
                description: Message is a human-readable explanation specific to this occurrence of the problem.
                example: parameter 'p' must be an integer
            name:
                type: string
                description: Name is the name of this class of errors.
                example: bad_request
            temporary:
                type: boolean
                description: Is the error temporary?
                example: false
            timeout:
                type: boolean
                description: Is the error a timeout?
                example: true
        description: listProduct_BadRequest_response_body result type (default view)
        example:
            fault: false
            id: 123abc
            message: parameter 'p' must be an integer
            name: bad_request
            temporary: false
            timeout: true
        required:
            - name
            - id
            - message
            - temporary
            - timeout
            - fault
...